requirements.txt
customtkinter
PyMuPDF
Pillow
PyPDF2
reportlab
